requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for testing
httpx==0.25.2
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Environment management
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Async utilities
aiofiles==23.2.1

# Optional: Database support (for future use)
# sqlalchemy==2.0.23
# alembic==1.13.0
# asyncpg==0.29.0  # for PostgreSQL
# aiomysql==0.2.0  # for MySQL

# Optional: Monitoring and logging
# structlog==23.2.0
# sentry-sdk==1.38.0

# Optional: Performance monitoring
# prometheus-client==0.19.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1